+++ packages/plugin/src/Integrations/EmailMarketing/MailerLite/MailerLiteIntegrationInterface.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Solspace\Freeform\Integrations\EmailMarketing\MailerLite;
+
+interface MailerLiteIntegrationInterface
+{
+    public function getApiKey(): string;
+}